Consolidate all constants files into /src/utils/const
