+++ src/js/App/CrossRequestNotifier/CrossRequestNotifier.tsx
@@ -1,7 +1,7 @@
 import React, { useReducer } from 'react';
 import Portal, { PortalNotificationConfig } from '@redhat-cloud-services/frontend-components-notifications/Portal';
 import { AlertVariant } from '@patternfly/react-core';
-import { ACCOUNT_REQUEST_TIMEOUT, ACTIVE_ACCOUNT_SWITCH_NOTIFICATION } from '../../consts';
+import { ACCOUNT_REQUEST_TIMEOUT, ACTIVE_ACCOUNT_SWITCH_NOTIFICATION } from '../../../utils/consts';
 import useAccessRequestNotifier from '../../../utils/useAccessRequestNotifier';
 import ChromeLink from '../Sidenav/Navigation/ChromeLink';
 import { useIntl } from 'react-intl';
