+++ src/js/App/GlobalFilter/globalFilterApi.ts
