+++ src/js/createCase.ts
@@ -4,7 +4,7 @@ import URI from 'urijs';
 const log = logger('createCase.js');
 
 import { getEnvDetails, getUrl } from '../utils/common';
-import { HYDRA_ENDPOINT } from './consts';
+import { HYDRA_ENDPOINT } from '../utils/consts';
 import { spinUpStore } from '../redux/redux-config';
 import { ChromeUser } from '@redhat-cloud-services/types';
 import { LibJWT } from './auth';
