+++ src/redux/chromeReducers.ts
@@ -1,6 +1,6 @@
 import { QuickStart } from '@patternfly/quickstarts';
 import { ChromeUser } from '@redhat-cloud-services/types';
-import { REQUESTS_COUNT, REQUESTS_DATA } from '../js/consts';
+import { REQUESTS_COUNT, REQUESTS_DATA } from '../utils/consts';
 import { NavItem } from '../js/types';
 import { generateRoutesList, highlightItems, isBeta, isFedRamp, levelArray } from '../utils/common';
 import { ThreeScaleError } from '../utils/responseInterceptors';
