+++ src/utils/useAccessRequestNotifier.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { useEffect, useRef } from 'react';
 import { batch, useDispatch, useSelector } from 'react-redux';
-import { REQUESTS_COUNT, REQUESTS_DATA } from '../js/consts';
+import { REQUESTS_COUNT, REQUESTS_DATA } from './consts';
 import { markAccessRequestNotification, updateAccessRequestsNotifications } from '../redux/actions';
 import { ReduxState } from '../redux/store';
